pyproject.toml

[project]
name = "walletpassesapp"
version = "0.1.0"
description = "This Desktop Application is to show the Classes and the Objects of the passes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flet[all]>=0.28.3",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "mysql-connector-python>=9.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "email-validator>=2.2.0",
    "requests>=2.32.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "qrcode[pil]>=7.4.0",
    "pillow>=10.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.2.1",
]
